requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
streamlit==1.35.0
openai==1.30.1
pydantic==2.7.3
pydantic-settings==2.3.0
python-dotenv==1.0.1
Pillow==10.3.0
httpx==0.27.0
requests==2.32.3
plotly==5.22.0
pandas==2.2.2